Solidify project vs program naming conventions
Naming conventions are currently inconsistent, due to evolution of documents
over time and intended audience.

P4XT should refer to the "program", as defined in the Abstract of the Phase 2
statement of work:
"Open Research Institute Multiplexing Transponder Program (P4XT) ..."

When using P4XT, don't add "program" or "project".  It's already in the definition.

Under P4XT, there are two co-projects, the transponder project (P4XDMT) and the modem
project (P4XM).  Note again, "project" is included in the definitions of both P4XDMT and
P4XM, so it is not necessary to add "project" after using either acronym.

Additionally, there will not be references to "ground" or "space" in the project section.  This
division is historical and can be confusing in the current context (i.e. the P4XDMT is now
targeted for both space and terrestrial use).